Create new utils file search.py Move model_keywords_search to utils/search.py
